requirements.txt
gradio
google-generativeai
requests
dateparser
geopy
huggingface_hub